+++ AdminSystem/OrderLineViewer.aspx
@@ -1,4 +1,4 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="5List.aspx.cs" Inherits="_1_List" %>
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="OrderLineViewer.aspx.cs" Inherits="_1Viewer" %>
 
 <!DOCTYPE html>
