+++ TestingOrders/tstOrders.cs
@@ -69,71 +69,6 @@ namespace TestingOrders
         }
 
         [TestMethod]
-        public void OrderLineIDOK()
-        {
-            //Create an instance of the class we will create
-            clsOrders Order = new clsOrders();
-            //Create some test data to assign to the property
-            int TestData = 1;
-            //Assign the data to the property
-            Order.OrderLineID = TestData;
-            //Test to see if the two values are the same
-            Assert.AreEqual(Order.OrderLineID, TestData);
-        }
-
-        [TestMethod]
-        public void OrderLineCheckoutOK()
-        {
-            //Create an instance of the class we will create
-            clsOrders Order = new clsOrders();
-            //Create some test data to assign to the property
-            Boolean TestData = true;
-            //Assign the data to the property
-            Order.OrderLineCheckout = TestData;
-            //Test to see if the two values are the same
-            Assert.AreEqual(Order.OrderLineCheckout, TestData);
-        }
-
-        [TestMethod]
-        public void CheckoutDateOK()
-        {
-            //Create an instance of the class we will create
-            clsOrders Order = new clsOrders();
-            //Create some test data to assign to the property
-            DateTime TestData = DateTime.Now.Date;
-            //Assign the data to the property
-            Order.CheckoutDate = TestData;
-            //Test to see if the two values are the same
-            Assert.AreEqual(Order.CheckoutDate, TestData);
-        }
-
-        [TestMethod]
-        public void OrderLineTotalOK()
-        {
-            //Create an instance of the class we will create
-            clsOrders Order = new clsOrders();
-            //Create some test data to assign to the property
-            int TestData = 2;
-            //Assign the data to the property
-            Order.OrderLineTotal = TestData;
-            //Test to see if the two values are the same
-            Assert.AreEqual(Order.OrderLineTotal, TestData);
-        }
-
-        [TestMethod]
-        public void TotalCostOK()
-        {
-            //Create an instance of the class we will create
-            clsOrders Order = new clsOrders();
-            //Create some test data to assign to the property
-            Double TestData = 12.50;
-            //Assign the data to the property
-            Order.TotalCost = TestData;
-            //Test to see if the two values are the same
-            Assert.AreEqual(Order.TotalCost, TestData);
-        }
-
-        [TestMethod]
         public void FindMethodOK()
         {
             clsOrders Order = new clsOrders();
@@ -197,79 +132,6 @@ namespace TestingOrders
             Int32 OrderID = 1;
             Found = Order.Find(OrderID);
             if (Order.OrderPlaced != true)
-            {
-                OK = false;
-            }
-            Assert.IsTrue(OK);
-        }
-
-        [TestMethod]
-        public void TestCheckoutDateFound()
-        {
-            clsOrders Order = new clsOrders();
-            Boolean Found = false;
-            Boolean OK = true;
-            Int32 OrderID = 1;
-            Found = Order.Find(OrderID);
-            if (Order.CheckoutDate != Convert.ToDateTime("18/02/2021"))
-            {
-                OK = false;
-            }
-            Assert.IsTrue(OK);
-        }
-
-        [TestMethod]
-        public void TestOrderLineIDFound()
-        {
-            clsOrders Order = new clsOrders();
-            Boolean Found = false;
-            Boolean OK = true;
-            Int32 OrderID = 2;
-            Found = Order.Find(OrderID);
-            if (Order.OrderLineID != 2)
-            {
-                OK = false;
-            }
-            Assert.IsTrue(OK);
-        }
-        [TestMethod]
-        public void testOrderLineCheckoutFound()
-        {
-            clsOrders Order = new clsOrders();
-            Boolean Found = false;
-            Boolean OK = true;
-            Int32 OrderID = 1;
-            Found = Order.Find(OrderID);
-            if (Order.OrderLineCheckout != true)
-            {
-                OK = false;
-            }
-            Assert.IsTrue(OK);
-        }
-
-        [TestMethod]
-        public void testOrderLineTotalFound()
-        {
-            clsOrders Order = new clsOrders();
-            Boolean Found = false;
-            Boolean OK = true;
-            Int32 OrderID = 1;
-            Found = Order.Find(OrderID);
-            if (Order.OrderLineTotal != 10)
-            {
-                OK = false;
-            }
-            Assert.IsTrue(OK);
-        }
-        [TestMethod]
-        public void testTotalCostFound()
-        {
-            clsOrders Order = new clsOrders();
-            Boolean Found = false;
-            Boolean OK = true;
-            Int32 OrderID = 1;
-            Found = Order.Find(OrderID);
-            if (Order.TotalCost != 9.99)
             {
                 OK = false;
             }
